Generalising scripts; error handling; number padding

+++ docker/builder/scale.bash
@@ -2,11 +2,16 @@
 
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
+PROJECT="${PROJECT:-lachesis}"
+
 # Cleanup
 rm -rf "$DIR/nodes" "$DIR/peers.json"
-nodes=$(docker ps -a --no-trunc --filter name='^/lachesis' --format '{{.Names}}')
-docker stop -f "$nodes"
-docker rm "$nodes"
+nodes=$(docker ps -a --no-trunc --filter name='^/'"$PROJECT" --format '{{.Names}}')
+
+if [ ! -z "$nodes" ]; then
+  docker stop -f "$nodes"
+  docker rm "$nodes"
+fi
 
 # Config
 n="${n:-1000}"
@@ -16,8 +21,8 @@ ip_range="$ip_start/$subnet"
 
 # Run
 batch-ethkey -dir nodes -network "$ip_start" -n "$n" > peers.json
-docker build --compress --squash --force-rm --tag lachesis "$DIR"
+docker build --compress --force-rm --tag "$PROJECT" "$DIR"
 "$DIR/network.bash" "$ip_range"
 "$DIR/spin_multi.bash" "$n"
 
-docker start $(docker ps -a --no-trunc --filter name='^/lachesis' --format '{{.Names}}')
+docker start $(docker ps -a --no-trunc --filter name='^/'"$PROJECT" --format '{{.Names}}')
